// src/client/Actions.ts
export enum Actions {
  SET_CONFIGURATION = 'SET_CONFIGURATION',
  BACKUP_SET_ID = 'BACKUP_SET_GIST_ID',
  BACKUP_SET_DESCRIPTION = 'BACKUP_SET_DESCRIPTION',
  BACKUP_SET_URL = 'BACKUP_SET_URL',
  INTERESTING_PROJECTS_FETCHING = 'INTERESTING_PROJECTS_FETCHING',
  INTERESTING_PROJECTS = 'INTERESTING_PROJECTS',
  INITIALISING = 'INITIALISING',
  NAVIGATED = 'NAVIGATED',
  FULL_SCREEN = 'FULL_SCREEN',
  REQUEST_FULL_SCREEN = 'REQUEST_FULL_SCREEN',
  NOTIFICATION = 'NOTIFICATION',
  NOTIFICATION_DISMISS = 'NOTIFICATION_DISMISS',
  SHOW_BUILD_TIME = 'SHOW_BUILD_TIME',
  SHOW_BROKEN_BUILD_TIME = 'SHOW_BROKEN_BUILD_TIME',
  SHOW_TRAY_NAME = 'SHOW_TRAY_NAME',
  PLAY_BROKEN_BUILD_SOUND_FX = 'PLAY_BROKEN_BUILD_SOUND_FX',
  BROKEN_BUILD_SOUND_FX = 'BROKEN_BUILD_SOUND_FX',
  REFRESH_TIME = 'REFRESH_TIME',
  SHOW_BUILD_LABEL = 'SHOW_BUILD_LABEL',
  REQUESTING_SYSTEM_NOTIFICATION_PERMISSION = 'REQUESTING_SYSTEM_NOTIFICATION_PERMISSION',
  SHOW_SYSTEM_NOTIFICATIONS = 'SHOW_SYSTEM_NOTIFICATIONS',
  SYSTEM_NOTIFICATIONS_PERMISSION_DENIED = 'SYSTEM_NOTIFICATIONS_PERMISSION_DENIED',
  SET_MAX_PROJECTS = 'SET_MAX_PROJECTS',
  CLICK_TO_SHOW_MENU = 'CLICK_TO_SHOW_MENU',
  SHOW_PROGNOSIS = 'SHOW_PROGNOSIS',
  MESSAGE_ADDED = 'MESSAGE_ADDED',
  MESSAGE_REMOVED = 'MESSAGE_REMOVED',
  TRAY_ADDED = 'TRAY_ADDED',
  HIGHLIGHT_TRAY = 'HIGHLIGHT_TRAY',
  ENCRYPTING_PASSWORD = 'ENCRYPTING_PASSWORD',
  PASSWORD_ENCRYPTED = 'PASSWORD_ENCRYPTED',
  PASSWORD_ENCRYPT_ERROR = 'PASSWORD_ENCRYPT_ERROR',
  ENCRYPTING_TOKEN = 'ENCRYPTING_TOKEN',
  TOKEN_ENCRYPTED = 'TOKEN_ENCRYPTED',
  TOKEN_ENCRYPT_ERROR = 'TOKEN_ENCRYPT_ERROR',
  REMOVE_TRAY = 'REMOVE_TRAY',
  PROJECTS_FETCHING = 'PROJECTS_FETCHING',
  PROJECTS_FETCHED = 'PROJECTS_FETCHED',
  PROJECTS_FETCH_ERROR = 'PROJECTS_FETCH_ERROR',
  SET_TRAY_NAME = 'SET_TRAY_NAME',
  SET_SERVER_TYPE = 'SET_SERVER_TYPE',
  SET_TRAY_USERNAME = 'SET_TRAY_USERNAME',
  SET_TRAY_AUTH_TYPE = 'SET_TRAY_AUTH_TYPE',
  SET_TRAY_URL = 'SET_TRAY_URL',
  SET_INCLUDE_NEW = 'SET_INCLUDE_NEW',
  SELECT_PROJECT = 'SELECT_PROJECT',
  ABORT_PENDING_REQUEST = 'ABORT_PENDING_REQUEST',
  NOOP = 'NOOP'
}